+++ ramls/streamingextractortask.json
@@ -4,7 +4,7 @@
   "type" : "object",
   "additionalProperties" : false,
   "properties" : {
-    "type" : {
+    "deserializeAs" : {
       "type" : "string",
       "enum" : [ "StreamingExtractorTask" ],
       "default" : "StreamingExtractorTask"
@@ -34,5 +34,5 @@
       "type" : "string"
     }
   },
-  "required" : [ "type", "name" ]
+  "required" : [ "deserializeAs", "name" ]
 }
