+++ src/main/java/org/folio/rest/model/repo/TriggerRepo.java
@@ -11,5 +11,6 @@ import org.springframework.http.HttpMethod;
 public interface TriggerRepo extends JpaRepository<Trigger, String> {
 
   public List<Trigger> findByMethod(HttpMethod method);
+  public List<Trigger> findByMethodAndDeserializeAs(HttpMethod method, String deserializeAs);
 
 }
